feat: add API v2 support with version-aware views
