+++ src/components/ApiKeyInput.vue
@@ -3,6 +3,14 @@
     <h2>API Configuration</h2>
     <form @submit.prevent="saveApiConfig">
       <div class="form-group">
+        <label for="apiVersion">Firecrawl API version:</label>
+        <select id="apiVersion" v-model="selectedVersion">
+          <option value="v1">Version 1</option>
+          <option value="v2">Version 2</option>
+        </select>
+        <small>Select the API version matching your backend deployment.</small>
+      </div>
+      <div class="form-group">
         <label for="apiKey">Firecrawl API Key:</label>
         <input
           id="apiKey"
@@ -18,43 +26,42 @@
       </div>
       <div class="form-group">
         <label for="baseUrl">Firecrawl API base URL (optional):</label>
-        <input
-          id="baseUrl"
-          v-model="baseUrl"
-          type="text"
-          placeholder="https://api.firecrawl.dev/v1"
-        />
-        <small> Leave blank to use the default URL. </small>
+        <input id="baseUrl" v-model="baseUrl" type="text" :placeholder="placeholderUrl" />
+        <small> Leave blank to use the default URL for the selected version. </small>
       </div>
       <button type="submit" class="primary-button">Save</button>
     </form>
     <div v-if="error" class="error-message">
       {{ error }}
     </div>
-    <div v-if="success" class="success-message">API Key saved successfully!</div>
+    <div v-if="success" class="success-message">API configuration saved successfully!</div>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, ref, onMounted } from 'vue';
+import { defineComponent, ref, onMounted, computed } from 'vue';
 import { refreshApiClients } from '@/plugins/api';
+import { getDefaultBaseUrl, getApiVersion, type ApiVersion } from '@/config/api';
+import { useApiVersion } from '@/stores/apiVersion';
 
 /**
- * Component allowing users to configure and store the Firecrawl API key and base URL.
+ * Component allowing users to configure and store the Firecrawl API key, version and base URL.
  * This component handles the persistence of these settings in local storage.
  *
  * @returns Component options for the API key input view.
  */
-
 export default defineComponent({
   name: 'ApiKeyInput',
   setup() {
+    const apiVersionState = useApiVersion();
+    const selectedVersion = ref<ApiVersion>(apiVersionState.value ?? getApiVersion());
     const apiKey = ref(localStorage.getItem('firecrawl_api_key') || '');
     const baseUrl = ref(localStorage.getItem('firecrawl_base_url') || '');
     const error = ref('');
     const success = ref(false);
 
-    // Automatically display if no key is configured
+    const placeholderUrl = computed(() => getDefaultBaseUrl(selectedVersion.value));
+
     onMounted(() => {
       if (!apiKey.value) {
         error.value = 'Please configure your API key to continue';
@@ -66,17 +73,19 @@ export default defineComponent({
         if (!apiKey.value) {
           throw new Error('Please enter a valid API key');
         }
-        // Save the API key
+
         localStorage.setItem('firecrawl_api_key', apiKey.value);
-        // Save the base URL (even if empty to override any previous value if needed)
-        localStorage.setItem('firecrawl_base_url', baseUrl.value);
+        if (baseUrl.value.trim()) {
+          localStorage.setItem('firecrawl_base_url', baseUrl.value);
+        } else {
+          localStorage.removeItem('firecrawl_base_url');
+        }
 
         success.value = true;
         error.value = '';
         setTimeout(() => (success.value = false), 3000);
 
-        // Update API clients dynamically without reloading the page
-        refreshApiClients(baseUrl.value, apiKey.value);
+        refreshApiClients(baseUrl.value, apiKey.value, selectedVersion.value);
       } catch (err) {
         error.value = err instanceof Error ? err.message : 'Unknown error';
         success.value = false;
@@ -89,6 +98,8 @@ export default defineComponent({
       error,
       success,
       saveApiConfig,
+      selectedVersion,
+      placeholderUrl,
     };
   },
 });
@@ -113,7 +124,8 @@ label {
   font-weight: bold;
 }
 
-input {
+input,
+select {
   width: 100%;
   padding: 8px;
   border: 1px solid #ddd;
