+++ src/plugins/api.ts
@@ -1,3 +1,5 @@
+import axios from 'axios';
+import type { AxiosInstance } from 'axios';
 import type { App } from 'vue';
 import {
   BillingApi,
@@ -7,35 +9,97 @@ import {
   ScrapingApi,
   SearchApi,
 } from '../api-client/index.js';
-import apiConfig, { getApiConfig, updateApiConfig } from '../config/api.js';
+import {
+  CrawlingApiV2,
+  ExtractionApiV2,
+  MappingApiV2,
+  ScrapingApiV2,
+  SearchApiV2,
+} from '../api-client-v2/index.js';
+import apiConfig, {
+  getApiConfig,
+  getApiKey,
+  getApiVersion,
+  getBaseUrl,
+  updateApiConfig,
+  type ApiVersion,
+} from '../config/api.js';
+import { setApiVersionState } from '@/stores/apiVersion.js';
 
 /**
- * Vue plugin responsible for registering and providing various Firecrawl API clients
- * to the Vue application instance.
- *
- * This plugin makes API clients accessible globally via `app.config.globalProperties.$api`
- * and through dependency injection via `app.provide('api', apis)`.
- *
- * @param app - The Vue application instance to which API clients will be registered.
+ * Combined client interface exposed through Vue injection.
  */
-interface FirecrawlApiClients {
-  billing: BillingApi;
-  crawling: CrawlingApi;
-  extraction: ExtractionApi;
-  mapping: MappingApi;
-  scraping: ScrapingApi;
-  search: SearchApi;
+export interface FirecrawlApiClients {
+  version: ApiVersion;
+  billing?: BillingApi;
+  crawling: CrawlingApi | CrawlingApiV2;
+  extraction: ExtractionApi | ExtractionApiV2;
+  mapping: MappingApi | MappingApiV2;
+  scraping: ScrapingApi | ScrapingApiV2;
+  search: SearchApi | SearchApiV2;
 }
 
 let appInstance: App | null = null;
 const apiClients: FirecrawlApiClients = {
-  billing: new BillingApi(apiConfig),
+  version: getApiVersion(),
+  billing: undefined,
   crawling: new CrawlingApi(apiConfig),
   extraction: new ExtractionApi(apiConfig),
   mapping: new MappingApi(apiConfig),
   scraping: new ScrapingApi(apiConfig),
   search: new SearchApi(apiConfig),
 };
+
+assignClients(apiClients, apiClients.version);
+setApiVersionState(apiClients.version);
+
+/**
+ * Create an Axios instance configured for the Firecrawl API.
+ *
+ * @param baseUrl - API base URL.
+ * @param apiKey - Optional API key to include in the Authorization header.
+ * @returns Configured Axios instance.
+ */
+function createAxiosInstance(baseUrl: string, apiKey: string): AxiosInstance {
+  return axios.create({
+    baseURL: baseUrl,
+    headers: {
+      'Content-Type': 'application/json',
+      ...(apiKey ? { Authorization: `Bearer ${apiKey}` } : {}),
+    },
+  });
+}
+
+/**
+ * Assign concrete API clients based on the selected API version.
+ *
+ * @param target - Mutable reference to the clients container.
+ * @param version - Active API version.
+ */
+function assignClients(target: FirecrawlApiClients, version: ApiVersion): void {
+  if (version === 'v1') {
+    target.version = 'v1';
+    target.billing = new BillingApi(getApiConfig());
+    target.crawling = new CrawlingApi(getApiConfig());
+    target.extraction = new ExtractionApi(getApiConfig());
+    target.mapping = new MappingApi(getApiConfig());
+    target.scraping = new ScrapingApi(getApiConfig());
+    target.search = new SearchApi(getApiConfig());
+    return;
+  }
+
+  const baseUrl = getBaseUrl(version);
+  const apiKey = getApiKey();
+  const httpClient = createAxiosInstance(baseUrl, apiKey);
+
+  target.version = 'v2';
+  target.billing = undefined;
+  target.crawling = new CrawlingApiV2(httpClient);
+  target.extraction = new ExtractionApiV2(httpClient);
+  target.mapping = new MappingApiV2(httpClient);
+  target.scraping = new ScrapingApiV2(httpClient);
+  target.search = new SearchApiV2(httpClient);
+}
 
 const apiPlugin = {
   install(app: App): void {
@@ -50,15 +114,13 @@ const apiPlugin = {
  *
  * @param baseUrl - Optional new base URL.
  * @param apiKey - Optional new API key.
+ * @param version - Optional API version override.
  */
-export function refreshApiClients(baseUrl?: string, apiKey?: string): void {
-  updateApiConfig(baseUrl, apiKey);
-  apiClients.billing = new BillingApi(getApiConfig());
-  apiClients.crawling = new CrawlingApi(getApiConfig());
-  apiClients.extraction = new ExtractionApi(getApiConfig());
-  apiClients.mapping = new MappingApi(getApiConfig());
-  apiClients.scraping = new ScrapingApi(getApiConfig());
-  apiClients.search = new SearchApi(getApiConfig());
+export function refreshApiClients(baseUrl?: string, apiKey?: string, version?: ApiVersion): void {
+  updateApiConfig(baseUrl, apiKey, version);
+  const resolvedVersion = getApiVersion();
+  assignClients(apiClients, resolvedVersion);
+  setApiVersionState(resolvedVersion);
 
   if (appInstance) {
     appInstance.config.globalProperties.$api = apiClients;
